Update archive.R
Adjust the logic of handling entries on the first day of the month - if it is the first day of the month, then we update the current and the previous months. This should solve the issue of some time tracks from the last day of the month not being accounted.